Add small delay after serial initialization
I noticed that if you change a bit the setup code (by for example including a digitalWrite), they board was not able to find the fingerprint Sensor. By adding a small delay after the serial initialization, it solved the issue.
It might be related to my setup, but I waste a lot of time wondering why it was not detecting the sensor.
Board: NodeMcu Esp8266 12e
Fingerprint sensor: Not an Adafruit, Aliexpress one.