Corrected action taken after succesfully storing the fingerprint.
A return statement after succesfully storing the fingerprint was missing. Added a return value of 0.
When the fingerprint is taken succesfully, the program should restart instead of continuing to take the fingerprint; amended.
The program prints dots to indicate it is waiting, removed a newline being printed after each of these in one case.